requirements.txt
discord.py
yt-dlp
ffmpeg-python
python-dotenv
pynacl
lyricsgenius
